Moduł 31 - Zapisywanie pokoju w bazie danych

+++ src/main/java/pl/clockworkjava/domain/reservation/Reservation.java
@@ -38,7 +38,7 @@ public class Reservation {
 
     public ReservationDTO getAsDTO() {
         return new ReservationDTO(this.id, this.from,
-                this.to, this.room.getId(),
+                this.to, (int)this.room.getId(),
                 this.room.getNumber(), this.guest.getId(),
                 this.guest.getFirstName() + " " + this.guest.getLastName());
     }
