+++ src/main/java/pl/clockworkjava/domain/room/dto/RoomDTO.java
@@ -2,13 +2,13 @@ package pl.clockworkjava.domain.room.dto;
 
 public class RoomDTO {
 
-    private int id;
+    private long id;
     private int number;
     private String beds;
     private int bedsCount;
     private int roomSize;
 
-    public RoomDTO(int id, int number, String beds, int bedsCount, int roomSize) {
+    public RoomDTO(long id, int number, String beds, int bedsCount, int roomSize) {
         this.id = id;
         this.number = number;
         this.beds = beds;
@@ -16,7 +16,7 @@ public class RoomDTO {
         this.roomSize = roomSize;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
